project_code/requirements.txt
anyio==3.6.2
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.2
docopt==0.6.2
fastapi==0.92.0
firebase-admin==6.1.0
gcloud==0.17.0
google-api-core==2.11.0
google-api-python-client==2.80.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.10.0
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.3
grpcio-status==1.51.3
h11==0.14.0
httplib2==0.21.0
idna==3.4
jws==0.1.3
msgpack==1.0.4
oauth2client==3.0.0
pipreqs==0.4.11
proto-plus==1.22.2
protobuf==4.22.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.4.3
pydantic==1.10.5
PyJWT==2.6.0
pyparsing==3.0.9
Pyrebase==3.0.27
python-jwt==2.0.1
requests==2.28.2
requests-toolbelt==0.7.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.4
starlette==0.25.0
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.14
uvicorn==0.20.0
yarg==0.1.9